chore: splittings, renamings
